+++ app/helpers/claims/show_helper.rb
@@ -7,5 +7,9 @@ module Claims
         content_tag(:h1, "Claim submitted", class: "govuk-panel__title", id: "submitted-title")
       end
     end
+
+    def shared_view_css_class_size(claim)
+      claim.has_ecp_policy? ? "l" : "xl"
+    end
   end
 end